requirements.txt
flask==2.3.3
rembg==2.0.50
Pillow==10.0.0
gunicorn==21.2.0
python-multipart==0.0.6
numpy==1.25.2
python-dotenv==1.0.0 